Release v0.2.1: Polish and Package Improvements
✨ Improvements:
- Fix clippy warning (unnecessary cast in downloader)
- Update Cargo.toml metadata (better description, keywords)
- Add progress bar for search index building
- Add Debian (.deb) package support

📦 New Package Formats:
- Ubuntu/Debian: .tar.gz + .deb
- Fedora/RHEL: .rpm
- Universal Linux: static .tar.gz
- Windows: .msi
- macOS: .tar.gz

🎨 UX Enhancements:
- Index building now shows real-time progress
- Better package descriptions
- Improved metadata for crates.io

See CHANGELOG.md for details.

+++ src/downloader.rs
@@ -369,7 +369,7 @@ impl WikiDownloader {
             }
 
             // Update progress (approximate based on buffer position)
-            pb.set_position(xml_reader.buffer_position() as u64);
+            pb.set_position(xml_reader.buffer_position());
             buf.clear();
         }
